requirements.txt
fastapi
uvicorn[standard]
pydantic<2
python-dotenv
langchain
langchain-google-genai
langchain_community
chromadb
unstructured
pypdf
streamlit